requirements.txt
torch
torch-geometric
deepchem
rdkit
biopython
fair-esm
scikit-learn
pandas
tqdm
PyTDC
lifelines
matplotlib
seaborn
scipy